working with api

+++ PrintingText/ViewModel/MainVM.cs
@@ -1,4 +1,5 @@
-﻿using Constructor.View;
+﻿using API;
+using Constructor.View;
 using Constructor.ViewModel;
 using System.ComponentModel;
 using System.IO;
@@ -19,12 +20,16 @@ namespace PrintingText.ViewModel
         private ITab selectTab;
         private PrintersSetting printersSetting = new PrintersSetting();
         private ConstructorTab constructorTab = new ConstructorTab();
+        private ApiTab apiTab = new ApiTab();
+
         private TemplateVM template;
+        private Document document;
 
         public MainVM()
         {
             SelectTab = constructorTab;
             Template = null;
+            document = new Document();
         }
 
         public ConstructorTab ConstructorTab
@@ -40,6 +45,22 @@ namespace PrintingText.ViewModel
             get
             {
                 return printersSetting;
+            }
+        }
+
+        public ApiTab ApiTab
+        {
+            get
+            {
+                return apiTab;
+            }
+        }
+
+        public Document Document
+        {
+            get
+            {
+                return document;
             }
         }
 
@@ -78,9 +99,20 @@ namespace PrintingText.ViewModel
             }
         }
 
-        public void ChangeTab(bool constructor)
+        public void ChangeTab(int constructor)
         {
-            SelectTab = constructor ? (ITab)constructorTab : (ITab)printersSetting;
+            if (constructor == 1)
+            {
+                SelectTab = constructorTab;
+            }
+            else if (constructor == 2)
+            {
+                SelectTab = printersSetting;
+            }
+            else if (constructor == 3)
+            {
+                SelectTab = apiTab;
+            }
         }
 
         public Grid RefreshPreviewArea(TableView TemplateArea)
@@ -114,16 +146,80 @@ namespace PrintingText.ViewModel
             writers.Write(fixedDocument, printersSetting.SelectPrinter.PrintTicket);
             doc.Close();
             package.Close();
-
             using (var fileStream = new StreamReader(printersSetting.path))
             {
-                using (var printStream = new PrintQueueStream
-                (printersSetting.SelectPrinter.PrintQueue, "Print Template", false, printersSetting.SelectPrinter.PrintTicket))
+                try
                 {
-                    fileStream.BaseStream.CopyTo(printStream);
+                    using (var printStream = new PrintQueueStream
+                    (printersSetting.SelectPrinter.PrintQueue, "Print Template", false, printersSetting.SelectPrinter.PrintTicket))
+                    {
+                        fileStream.BaseStream.CopyTo(printStream);
+                    }
+                }
+                catch
+                {
+                    
                 }
             }
             File.Delete(printersSetting.path);    
+        }
+        #endregion
+
+        #region Сохранение файла
+        public void SaveFile(TableView TemplateArea)
+        {
+            if (PrintersSetting.IsSaveToFile && PrintersSetting.IsSaveToPNG)
+            {
+                //var place = Document.Place(printersSetting.Page, TemplateArea);
+                RenderTargetBitmap renderTargetBitmap = new RenderTargetBitmap((int)TemplateArea.ActualWidth, (int)TemplateArea.ActualHeight, 96, 96, PixelFormats.Pbgra32);
+                renderTargetBitmap.Render(TemplateArea);
+                PngBitmapEncoder pngImage = new PngBitmapEncoder();
+                pngImage.Frames.Add(BitmapFrame.Create(renderTargetBitmap));
+                using (Stream fileStream = File.Create("Template\\" + "image.png"))
+                {
+                    pngImage.Save(fileStream);
+                }
+            }
+            else if (PrintersSetting.IsSaveToFile && PrintersSetting.IsSaveToPDF)
+            {
+                var fixedPage = new FixedPage();
+                if (printersSetting.Page.IsPortrait)
+                {
+                    fixedPage.Width = printersSetting.Page.Width;
+                    fixedPage.Height = printersSetting.Page.Height;
+                }
+                else
+                {
+                    fixedPage.Width = printersSetting.Page.Height;
+                    fixedPage.Height = printersSetting.Page.Width;
+                }
+                var place = Document.Place(printersSetting.Page, TemplateArea);
+                fixedPage.Children.Add(place);
+                var pageContent = new PageContent();
+                ((IAddChild)pageContent).AddChild(fixedPage);
+                var package = Package.Open(printersSetting.path, FileMode.Create);
+                var doc = new XpsDocument(package);
+                var writers = XpsDocument.CreateXpsDocumentWriter(doc);
+                var fixedDocument = new FixedDocument();
+                fixedDocument.Pages.Add(pageContent);
+                writers.Write(fixedDocument, printersSetting.SelectPrinter.PrintTicket);
+                doc.Close();
+                package.Close();
+                using (var fileStream = new StreamReader(printersSetting.path))
+                {
+                    try
+                    {
+                        using (var printStream = new PrintQueueStream
+                        (printersSetting.SelectPrinter.PrintQueue, "Print Template", false, printersSetting.SelectPrinter.PrintTicket))
+                        {
+                            fileStream.BaseStream.CopyTo(printStream);
+                        }
+                    }
+                    catch
+                    { }
+                }
+                File.Delete(printersSetting.path);
+            }
         }
         #endregion
 
@@ -135,15 +231,22 @@ namespace PrintingText.ViewModel
             {
                 Template = (TemplateVM)jsonFormatter.ReadObject(fs);
             }
+            if (ApiTab.Dossier.SelectField != null)
+            {
+                Template.FillInTheData(ApiTab.Dossier.SelectField);
+            }
         }
         #endregion
 
         #region Удаление шаблона
         public void DeleteTemplate()
         {
-            File.Delete(constructorTab.SelectedFiles.Url);
-            constructorTab.ReloadingCollectionFiles();
-            Template = null;
+            if (constructorTab.SelectedFiles.Url != null )
+            {
+                File.Delete(constructorTab.SelectedFiles.Url);
+                constructorTab.ReloadingCollectionFiles();
+                Template = null;
+            }
         }
         #endregion
     }
